some ui and ux changes

+++ frontend/src/pages/Admin/AdminPages/FeaturedAdmin.jsx
@@ -62,12 +62,26 @@ const FeaturedAdmin = () => {
   const searchData = (e) => {
     setValue(e.target.value)
   }
-  const handleDelete = async (id) => {
-    await axios.delete(`http://localhost:3070/featured/${id}`)
-    getData();
-    window.alert("Are you sure you want to delete this item?");
+  // const handleDelete = async (id) => {
+  //   await axios.delete(`http://localhost:3070/featured/${id}`)
+  //   getData();
+  //   window.alert("Are you sure you want to delete this item?");
 
-  }
+  // }
+
+  const handleDelete = async (id) => {
+    const confirmed = window.confirm("Are you sure you want to delete this item?");
+
+    if (confirmed) {
+      try {
+        await axios.delete(`http://localhost:3070/featured/${id}`);
+        getData();
+      } catch (error) {
+        console.error(error);
+      }
+    }
+  };
+
 
   const handleChange = (e) => {
     setState({ ...state, [e.target.name]: e.target.value })
